eliminate MutableSubSequence associated type as dave suggested
