+++ src/algo/mutable_collection_ext/mod.rs
@@ -3,12 +3,12 @@
 
 use crate::algo::reorderable_collection_ext::ReorderableCollectionExt;
 use crate::iterators::MutableCollectionIter;
-use crate::{MutableCollection, UnsafeMutableSubSequence};
+use crate::MutableCollection;
 
 /// Algorithms for `MutableCollection`.
 pub trait MutableCollectionExt: MutableCollection
 where
-    Self::MutableSubSequence: UnsafeMutableSubSequence,
+    Self::SubSequence: MutableCollection,
 {
     /*-----------------Iteration Algorithms-----------------*/
 
@@ -38,9 +38,7 @@ where
     }
 
     /// Returns an iterator to iterate over mutable element refs in collection.
-    fn iter_mut(
-        &mut self,
-    ) -> MutableCollectionIter<'_, Self::MutableSubSequence> {
+    fn iter_mut(&mut self) -> MutableCollectionIter<'_, Self::SubSequence> {
         MutableCollectionIter::new(self.full_mut())
     }
 }
@@ -48,6 +46,6 @@ where
 impl<R> MutableCollectionExt for R
 where
     R: MutableCollection + ?Sized,
-    R::MutableSubSequence: UnsafeMutableSubSequence,
+    R::SubSequence: MutableCollection,
 {
 }
