+++ src/algo/random_access_collection_ext/mod.rs
@@ -1,16 +1,13 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2025 Rishabh Dwivedi (rishabhdwivedi17@gmail.com)
 
-use crate::{
-    RandomAccessCollection, ReorderableCollection, UnsafeReorderableSubSequence,
-};
+use crate::{RandomAccessCollection, ReorderableCollection};
 mod sort;
 
 /// Algorithms for `RandomAccessCollection`.
 pub trait RandomAccessCollectionExt: RandomAccessCollection
 where
     Self::SubSequence: RandomAccessCollection,
-    Self::MutableSubSequence: RandomAccessCollection,
 {
     /*-----------------Sorting Algorithms-----------------*/
 
@@ -36,7 +33,7 @@ where
     fn sort_unstable_by<Compare>(&mut self, are_in_increasing_order: Compare)
     where
         Self: ReorderableCollection,
-        Self::MutableSubSequence: UnsafeReorderableSubSequence,
+        Self::SubSequence: ReorderableCollection,
         Compare: Fn(&Self::Element, &Self::Element) -> bool + Clone,
     {
         sort::sort_unstable_by(self, are_in_increasing_order);
@@ -61,7 +58,7 @@ where
     fn sort_unstable(&mut self)
     where
         Self: ReorderableCollection,
-        Self::MutableSubSequence: UnsafeReorderableSubSequence,
+        Self::SubSequence: ReorderableCollection,
         Self::Element: Ord,
     {
         self.sort_unstable_by(|x, y| x < y)
@@ -72,6 +69,5 @@ impl<R> RandomAccessCollectionExt for R
 where
     R: RandomAccessCollection + ?Sized,
     R::SubSequence: RandomAccessCollection,
-    R::MutableSubSequence: RandomAccessCollection,
 {
 }
