+++ src/std_impl/array/mod.rs
@@ -3,9 +3,7 @@
 
 pub mod array_impl;
 pub mod array_slice;
-pub mod array_slice_mut;
 pub mod slice_impl;
 pub mod vec_impl;
 
 pub use array_slice::*;
-pub use array_slice_mut::*;
