+++ src/std_impl/mod.rs
@@ -6,3 +6,4 @@ pub mod range;
 pub mod string;
 
 pub use array::*;
+pub use string::*;
